Task List Implementation
Started the implementation of the task list that will be popping up from the right side of the screen alongside its form for adding a task.
Issues that I found where of the popping up of the form, as it wasn't able to be interacted with ,furthermore it's styling needs to be configured, due to not fitting in one single screen without scrolling.